pyproject.toml
[project]
name = "wheather-api"
version = "0.1.0"
description = "Api to get weather data"
authors = [
    {name = "KhudyakovGleb",email = "glebkhudyakov@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "fastapi-cli (>=0.0.7,<0.0.8)",
    "uuid (>=1.30,<2.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
flake8 = "^7.1.2"
black = "^25.1.0"
pylint = "^3.3.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
